+++ packages/11ty/_includes/components/lightbox/slides.js
@@ -46,15 +46,15 @@ module.exports = function(eleventyConfig) {
             return figureVideoElement(figure)
           case mediaType === 'image':
           default:
-            return figureImageElement(figure, { preset: 'zoom' })
+            return figureImageElement(figure, { preset: 'zoom', interactive: true })
         }
       }
 
       const labelSpan = label
-        ? html`<span class="q-lightbox-slides__caption-label">${label}</span>`
+        ? html`<span class="q-lightbox-slides__caption-label">${markdownify(label)}</span>`
         : ''
       const captionAndCreditSpan = caption || credit
-        ? html`<span class="q-lightbox-slides__caption-content">${caption ? markdownify(caption) : ''} ${credit ? credit : ''}</span>`
+        ? html`<span class="q-lightbox-slides__caption-content">${caption ? markdownify(caption) : ''} ${credit ? markdownify(credit) : ''}</span>`
         : ''
       const captionElement = labelSpan.length || captionAndCreditSpan.length
         ? html`