+++ packages/cli/CHANGELOG.md
@@ -18,6 +18,18 @@ Changelog entries are classified using the following labels:
 
 - `Removed`: for deprecated features removed in this release
 
+## [1.0.0-rc.10]
+
+### Added
+- Support semantic version ranges for `quire-11ty` in quire starter projects and the `--quire-version` option of the `quire new` command, and use the latest compatible `quire-11ty` version when creating and installing a new project.
+
+## [1.0.0-rc.9]
+
+### Added
+- Update old `.quire` files  when running `info` command on versions prior to `1.0.0-rc.8`
+- Include `quire-cli` version used to generate project in `info` output
+- Adds headings to `info` output
+
 ## [1.0.0-rc.8]
 
 ### Added